Extract TideField into its own module
Break the propagator-owned tidal fold-in field out of
dnumerical_orbit_propagator.rs into src/propagators/tide_field.rs. The
propagator file had grown large and the TideField struct is
self-contained, so a dedicated private module keeps it readable and
reviewable.

Move the struct, its impl, and its four TideField-specific unit tests
(fold-in linearity, epoch-cache idempotency, point-mass delta-only,
pole-tide-requires-EOP) verbatim; propagation-level tests stay in place.
Register `mod tide_field;` as a private module and expose the type as
`pub(crate)`. No behavior change.
